Add use of the new isatab2json parser when possible
Some of the MTBLS studies fail to parse with the new  isatab2json().parser() due to some parameters missing in the i_*.txt file. If these cases the old version parser is used, although it takes much more time to complete. Note that In both cases the parsing is done without validation (validate_first=False).