Add Get and Post methods to directly access and update the title of a study, without using ISA-Tools
The current parser for ISA-Tab files is taking too much time to read the files and build a python object. This is not convenient for small changes like simply update the title of a study in the i_investigation.txt file, that has not other side effects in the s_*.txt or a_*.txt files. So the approach taken is to directly read and write the i_investigation.txt file, until a new parser and utility functions are developed in the ISA-API.