Add context to properly handle envelop on JSON responses
Envelope is an extra json tag with the name of the object returned when serializing a part of the study. This tag, single or plural, should only be present when requesting for the particular object (i.e. Sample) in a study, but not when is part of another object (i.e. a Sample in a Process).