Add POST method for Study Samples
A new Sample can be added to the Study. The input must be a valid ISA-JSON Sample file, including the Source(s) the Sample derives from. It is assumed a 'Sample collection' is used to produce the sample. A new 'Sample collection' protocol is created if not already present in the Study.